docs: increase the privilege expiration delay to six months
There are a number of cases where a contributor with escalated
privileges (for i18n or development) can be inactive for a few
months. It does not necessarily mean they are no longer motivated or
that they lost contact with SecureDrop for so long that they would not
be comfortable using their privileges.

However, if six months passed without activity we can safely assume it
would not bother them to re-apply again. And establishing this
revokation period is a non-controversial way to ensure all persons
with privileges are currently active. This is also the period
currently used by the tor project in a similar case.