source: genkey runs in a thread and must have its own session
The session available from Source.query belongs to the main thread
and is bound to a Flask app context. However, since async_genkey runs
in a separate thread, it is entirely possible that there is no active
app context when it tries to access the session and it will fail with
an error like:

    No application found. Either work inside a view function or
    push an application context.
    See http://flask-sqlalchemy.pocoo.org/contexts/

In addition the query from the thread races against the queries from
the main thread and it may lead to failure or corruption.

Since there is no need for async_genkey to work in an app_context, we
create a new engine and a new session for the duration of the thread.
Writing to SQLite from multiple processes or threads is protected by
locks (http://www.sqlite.org/wal.html)